+++ packages/designer/src/document/document-model.ts
@@ -50,6 +50,11 @@ export class DocumentModel {
   private rootNodeVisitorMap: { [visitorName: string]: any } = {};
 
   /**
+   * @deprecated
+   */
+  private _addons: { [key: string]: { exportData: () => any; isProp: boolean;} } = {};
+
+  /**
    * 模拟器
    */
   get simulator(): ISimulatorHost | null {
@@ -173,7 +178,7 @@ export class DocumentModel {
       node = this.getNode(schema.id);
       if (node && node.componentName === schema.componentName) {
         if (node.parent) {
-          node.internalSetParent(null);
+          node.internalSetParent(null, false);
           // will move to another position
           // todo: this.activeNodes?.push(node);
         }
@@ -512,6 +517,32 @@ export class DocumentModel {
     this.emitter.emit('lowcode_engine_renderer_ready', renderer);
   }
 
+  /**
+   * @deprecated
+   */
+  getAddonData(name: string) {
+    const addon = this._addons[name];
+    return addon?.exportData();
+  }
+
+  /**
+   * @deprecated
+   */
+  registerAddon(name: string, exportData: any) {
+    if (['id', 'params', 'layout'].indexOf(name) > -1) {
+      throw new Error('addon name cannot be id, params, layout');
+    }
+    const i = this._addons?.findIndex((item) => item.name === name);
+    if (i > -1) {
+      this._addons?.splice(i, 1);
+    }
+    this._addons?.push({
+      exportData,
+      name,
+    });
+  }
+
+
   acceptRootNodeVisitor(
     visitorName: string = 'default',
     visitorFn: (node: RootNode) => any ) {
@@ -545,6 +576,13 @@ export class DocumentModel {
     return () => {
       this.emitter.removeListener('nodedestroy', func);
     };
+  }
+
+  /**
+   * @deprecated
+   */
+  refresh() {
+    console.warn('refresh method is deprecated');
   }
 }