+++ packages/editor-skeleton/src/transducers/addon-combine.ts
@@ -132,10 +132,11 @@ export default function(metadata: TransformedComponentMetadata): TransformedComp
     let l = propsGroup.length;
     while (l-- > 0) {
       const item = propsGroup[l];
-      if (item.type === 'group' && (item.title === '高级' || item.title?.label === '高级')) {
-        advanceGroup = item.items || [];
-        propsGroup.splice(l, 1);
-      } else if (item.name === '__style__' || item.name === 'containerStyle' || item.name === 'pageStyle') {
+      // if (item.type === 'group' && (item.title === '高级' || item.title?.label === '高级')) {
+      //   advanceGroup = item.items || [];
+      //   propsGroup.splice(l, 1);
+      // } 
+      if (item.name === '__style__' || item.name === 'containerStyle' || item.name === 'pageStyle') {
         propsGroup.splice(l, 1);
         stylesGroup.push(item);
         if (item.extraProps?.defaultCollapsed && item.name !== 'containerStyle') {
@@ -212,6 +213,9 @@ export default function(metadata: TransformedComponentMetadata): TransformedComp
         }, {
           componentName: 'VariableSetter'
         }],
+        extraProps: {
+          display: 'block',
+        },
       });
     }
     if (supports.loop !== false) {
@@ -252,18 +256,28 @@ export default function(metadata: TransformedComponentMetadata): TransformedComp
               }
             },
           },
-          {
-            name: 'key',
-            title: '循环 Key',
-            setter: [{
-              componentName: 'StringSetter',
-            }, {
-              componentName: 'VariableSetter'
-            }],
-          },
         ],
+        extraProps: {
+          display: 'accordion',
+        },
       })
     }
+    advanceGroup.push({
+      name: 'key',
+      title: {
+        label: '渲染唯一标识（key）',
+        tip: '搭配「条件渲染」或「循环渲染」时使用，和 react 组件中的 key 原理相同，点击查看帮助',
+        docUrl: 'https://yuque.antfin-inc.com/legao/help3.0/ca5in7',
+      },
+      setter: [{
+        componentName: 'StringSetter',
+      }, {
+        componentName: 'VariableSetter'
+      }],
+      extraProps: {
+        display: 'block',
+      },
+    },)
   }
   if (advanceGroup.length > 0) {
     combined.push({